refactor: Split large files into focused modules
Split harbor-api/src/routes/management.rs into:
- management/mod.rs: Module definitions and router construction
- management/auth.rs: Authentication routes and extractors
- management/users.rs: User CRUD operations
- management/cache.rs: Cache statistics and cleanup
- management/config.rs: Configuration management
- management/types.rs: Request/Response DTOs

Split harbor-db/src/repository.rs into:
- repository/mod.rs: Database struct and migrations
- repository/cache.rs: Cache entry operations
- repository/users.rs: User operations
- repository/sessions.rs: Upload session operations
- repository/config.rs: Configuration operations

Each module is now under 200 lines and focused on a single responsibility.
All functionality maintained, tests pass, no breaking changes.
