+++ crates/harbor-db/src/repository/mod.rs
@@ -1,0 +1,125 @@
+//! Database repository implementation
+
+use sqlx::SqlitePool;
+use tracing::info;
+
+use crate::error::DbError;
+
+// Submodules
+mod cache;
+mod config;
+mod sessions;
+mod users;
+
+// Re-export CacheStats
+pub use cache::CacheStats;
+
+/// Database connection and operations
+#[derive(Clone)]
+pub struct Database {
+    pool: SqlitePool,
+}
+
+impl Database {
+    /// Create a new database connection
+    pub async fn new(database_url: &str) -> Result<Self, DbError> {
+        info!("Connecting to database: {}", database_url);
+
+        let pool = SqlitePool::connect(database_url).await?;
+        let db = Self { pool };
+        db.run_migrations().await?;
+        Ok(db)
+    }
+
+    /// Get the underlying pool for advanced usage
+    pub fn pool(&self) -> &SqlitePool {
+        &self.pool
+    }
+
+    /// Run database migrations
+    async fn run_migrations(&self) -> Result<(), DbError> {
+        info!("Running database migrations");
+
+        // Create tables if they don't exist
+        sqlx::query(
+            r#"
+            CREATE TABLE IF NOT EXISTS cache_entries (
+                id INTEGER PRIMARY KEY AUTOINCREMENT,
+                entry_type TEXT NOT NULL,
+                repository TEXT,
+                reference TEXT,
+                digest TEXT NOT NULL UNIQUE,
+                content_type TEXT NOT NULL,
+                size INTEGER NOT NULL,
+                created_at TEXT NOT NULL,
+                last_accessed_at TEXT NOT NULL,
+                access_count INTEGER DEFAULT 1,
+                storage_path TEXT NOT NULL
+            )
+            "#,
+        )
+        .execute(&self.pool)
+        .await?;
+
+        sqlx::query(
+            r#"
+            CREATE INDEX IF NOT EXISTS idx_cache_entries_digest ON cache_entries(digest)
+            "#,
+        )
+        .execute(&self.pool)
+        .await?;
+
+        sqlx::query(
+            r#"
+            CREATE INDEX IF NOT EXISTS idx_cache_entries_last_accessed ON cache_entries(last_accessed_at)
+            "#,
+        )
+        .execute(&self.pool)
+        .await?;
+
+        sqlx::query(
+            r#"
+            CREATE TABLE IF NOT EXISTS users (
+                id INTEGER PRIMARY KEY AUTOINCREMENT,
+                username TEXT NOT NULL UNIQUE,
+                password_hash TEXT NOT NULL,
+                role TEXT NOT NULL,
+                created_at TEXT NOT NULL,
+                updated_at TEXT NOT NULL
+            )
+            "#,
+        )
+        .execute(&self.pool)
+        .await?;
+
+        sqlx::query(
+            r#"
+            CREATE TABLE IF NOT EXISTS config (
+                key TEXT PRIMARY KEY,
+                value TEXT NOT NULL,
+                updated_at TEXT NOT NULL
+            )
+            "#,
+        )
+        .execute(&self.pool)
+        .await?;
+
+        sqlx::query(
+            r#"
+            CREATE TABLE IF NOT EXISTS upload_sessions (
+                id TEXT PRIMARY KEY,
+                repository TEXT NOT NULL,
+                started_at TEXT NOT NULL,
+                last_chunk_at TEXT NOT NULL,
+                bytes_received INTEGER DEFAULT 0,
+                temp_path TEXT NOT NULL
+            )
+            "#,
+        )
+        .execute(&self.pool)
+        .await?;
+
+        info!("Database migrations completed");
+        Ok(())
+    }
+}
